Bubble_sort.go
This Go program (bubble_sort.go) provides an implementation of the Bubble Sort algorithm. Bubble Sort is a simple sorting algorithm that repeatedly steps through the list, compares adjacent elements, and swaps them if they are in the wrong order. The pass through the list is repeated until no swaps are needed, which indicates that the list is sorted. This code is a basic example of sorting in Go and can be used as a reference for understanding the Bubble Sort algorithm in the Go programming language.